Custom build script tag with JS
the privacy plugin automatically downloads external sources and hosts them on our site - which is mostly great, but needed an exception for our self hosted plausible script.

there is the option to exclude_assets for download, but that is only available to sponsors. As a result, this is a working workaround

https://github.com/squidfunk/mkdocs-material/pull/6502